Fixed a critical bug where we were not clearing the vector entry when reading the powersupply, that's why we alway get the first value, which is there fPsVector[0] for Top Powersupply and fPsVector[1] or Bottom supply.
Now correctly clearing the vector before filling it again with Powersupply read values,
and as a result of this the vector size is always 2, and hence we get updated values
on each fetch.

Also fix a minor bug in PlotVmonImon.cpp, where previously we were wrongle setting vmon to [i][i]
which ideally should be [i][j].

Also added a flag ** USE_TIME_ON_AXIS **, which when swithed on shows time on X axis
and when swithched OFF, then it will just show continuously increasing sequence on
X axis. Useful when we just wanted to see the fluctuation in Voltage or current for
debugging purpose.

Hopefully this fixes all the bugs discovered so far. This needs to be updated in the lab.